made getMainAddress public
