+++ packages/engine/paima-sm/package.json
@@ -15,7 +15,6 @@
   "author": "Paima Studios",
   "dependencies": {
     "@sinclair/typebox": "^0.31.28",
-    "assert-never": "^1.2.1",
-    "lru-cache": "^10.1.0"
+    "assert-never": "^1.2.1"
   }
 }
